// src/api/types.gen.ts
export type EventShortEntry = {
    id: string;
    name: string;
    description?: string | null;
    eventDateStart: string;
    eventDateEnd?: string | null;
    isCompleted: boolean;
    imageUri?: string | null;
};
// ...
export type GetParticipantEntry = {
    id: string;
    user: GetUserEntry;
    event: EventShortEntry;
    purchases?: Array<PurchaseShortEntry> | null;
    purchaseUsages?: Array<PurchaseUsageShortEntry> | null;
};

export type GetPurchaseEntry = {
    id: string;
    name: string;
    cost: number;
    isComplete: boolean;
    eventId: string;
    purchaseUsages?: Array<PurchaseUsageShortEntry> | null;
    participant?: ParticipantShortEntry;
};

export type GetPurchaseEntryResponse = {
    data?: Array<GetPurchaseEntry> | null;
    totalPages?: number;
    total?: number;
};

export type GetUserEntry = {
    id?: string | null;
    name?: string | null;
    email?: string | null;
    role?: string | null;
    firstName?: string | null;
    middleName?: string | null;
    lastName?: string | null;
    phone?: string | null;
    imageUri?: string | null;
};

export type LoginModel = {
    username: string;
    password: string;
};

export type MediaShortEntry = {
    id?: string;
    fileType?: FileType;
    fileName?: string | null;
    mediaType?: MediaType;
    link?: string | null;
};

export type MediaType = 'Image' | 'Video' | 'Audio' | 'Document' | 'Link' | 'Unknown';

export type ParticipantShortEntry = {
    id: string;
    name?: string | null;
    imageUri?: string | null;
};

export type ParticipantShortEntryResponse = {
    data?: Array<ParticipantShortEntry> | null;
    totalPages?: number;
    total?: number;
};

export type PurchaseShortEntry = {
    id: string;
    eventId: string;
    name: string;
    cost: number;
    isComplete: boolean;
};

export type PurchaseUsageShortEntry = {
    purchaseUsageId: string;
    purchaseId?: string;
    participantId?: string;
    amount?: number;
};
// ...
export type UpsertParticipantEntry = {
    userId: string;
};